requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
paho-mqtt>=1.6.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
ruff>=0.1.0
